+++ src/controllers/usercontroller.ts
@@ -2,17 +2,22 @@ import { Request, Response, NextFunction } from 'express';
 import { User } from '../models/user';
 import { UserService } from '../apis/services/userservice';
 import { ValidacaoUser } from '../utils/validacoes/validacoesuser';
+import { JsonWebToken } from '../utils/webtoken/jsonwebtoken';
 
 //melhorar validacoes nos metodos
 export class UserController {
 
     async index(req: Request, res: Response) {
-        try {
-            const userservice = new UserService();
-            const results = await userservice.index()
-            return res.json(results)
-        } catch (err) {
-            return res.status(401).json("Ocorreu algum erro!")
+        if (new JsonWebToken().verificaToken(req.headers.authorization)) {
+            try {
+                const userservice = new UserService();
+                const results = await userservice.index()
+                return res.json(results)
+            } catch (err) {
+                return res.status(400).json("Ocorreu algum erro!")
+            }
+        } else {
+            return res.status(401).json("Você não tem permissão para esta rota !")
         }
     }
 
@@ -24,28 +29,38 @@ export class UserController {
     }
 
     async update(req: Request, res: Response, next: NextFunction) {
-        const body = req.body as User
-        const id = req.params.id
-        try {
-            const user = new User()
-            var result = await user.UpdateUser(body.name, body.password, Number(id), body.role,
-                body.numerotelefone, body.cpf, body.datanascimento)
-            if (result) {
-                return res.status(201).json("Alterado com sucesso !")
-            } else {
-                return res.status(400).json("Ocorreu algum erro ao alterar!")
+        if (new JsonWebToken().verificaToken(req.headers.authorization)) {
+            const body = req.body as User
+            const id = req.params.id
+            try {
+                const user = new User()
+                var result = await user.UpdateUser(body.name, body.password, Number(id), body.role,
+                    body.numerotelefone, body.cpf, body.datanascimento)
+                if (result) {
+                    return res.status(201).json("Alterado com sucesso !")
+                } else {
+                    return res.status(400).json("Ocorreu algum erro ao alterar!")
+                }
             }
+            catch (err) {
+                next(err)
+            }
+        } else {
+            return res.status(401).json("Você não tem permissão para esta rota !")
         }
-        catch (err) {
-            next(err)
-        }
+
     }
 
     async delete(req: Request, res: Response, next: NextFunction) {
-        const id = req.params.id
-        const userservice = new UserService();
-        var respostas = await userservice.delete(id)
-        return res.status(respostas.status).json(respostas.resposta)
+        if (new JsonWebToken().verificaToken(req.headers.authorization)) {
+            const id = req.params.id
+            const userservice = new UserService();
+            var respostas = await userservice.delete(id)
+            return res.status(respostas.status).json(respostas.resposta)
+        } else {
+            return res.status(401).json("Você não tem permissão para esta rota !")
+        }
+
     }
 
     async login(req: Request, res: Response, next: NextFunction) {
